Upgrade dependencies and change <Button> API due to changes in TypeScript
The improved support for generics means we now can't use the Type key for the ButtonProps, so we now use Kind instead.

We also upgraded the react router dependency, but this is not really a breaking change, it's just due to the dependency resolution that they had to bump the major.